+++ src/devices/laserFromPointCloud/laserFromPointCloud.h
@@ -53,8 +53,8 @@ protected:
     PolyDriver m_tc_driver;
     IFrameTransform* m_iTc = nullptr;
 
-    int m_depth_width = 0;
-    int m_depth_height = 0;
+    size_t m_depth_width = 0;
+    size_t m_depth_height = 0;
     yarp::os::Property m_propIntrinsics;
     yarp::sig::IntrinsicParams m_intrinsics;
     yarp::sig::ImageOf<float> m_depth_image;
